feat: splitted test file in multiple test files
