+++ test/helpers/evm-utils.ts
@@ -1,4 +1,3 @@
-import HRE from 'hardhat';
 import { HardhatRuntimeEnvironment } from 'hardhat/types';
 
 export async function increaseTime(
@@ -9,10 +8,16 @@ export async function increaseTime(
   await hre.ethers.provider.send('evm_mine', []);
 }
 
-export async function evmSnapshot(): Promise<void> {
-  await HRE.ethers.provider.send('evm_snapshot', []);
+export async function evmSnapshot(
+  hre: HardhatRuntimeEnvironment
+): Promise<string> {
+  const snapshotId = hre.ethers.provider.send('evm_snapshot', []);
+  return snapshotId;
 }
 
-export async function evmRevert(id: string): Promise<void> {
-  await HRE.ethers.provider.send('evm_revert', [id]);
+export async function evmRevert(
+  hre: HardhatRuntimeEnvironment,
+  snapshotId: string
+): Promise<void> {
+  await hre.ethers.provider.send('evm_revert', [snapshotId]);
 }
